MSVC 2015 support

+++ example/rng/include/rng_dist.hpp
@@ -154,7 +154,8 @@ inline double rng_dist_chi2(
     double p = 0;
     for (std::size_t i = 0; i != partition.size(); ++i)
         p += (count[i] - e) * (count[i] - e) / e;
-    boost::math::chi_squared_distribution<double> chi2(partition.size() - 1);
+    boost::math::chi_squared_distribution<double> chi2(
+            static_cast<double>(partition.size() - 1));
 
     return boost::math::cdf(chi2, p);
 }
